+++ Source/SteamNotifier/App.config
