+++ Source/SteamNotifier/Properties/AssemblyInfo.cs
