+++ Source/SteamNotifier/Properties/Resources.resx
