+++ Source/SteamNotifier/Properties/Settings.settings
