+++ Source/SteamNotifier/RegistryMonitor.cs
