+++ Source/SteamNotifier/Resources/icon-bg.ico
