+++ Source/SteamNotifier/Settings.cs
