+++ Source/SteamNotifier/SteamNotifier.csproj
