+++ Source/SteamNotifier/packages.config
