+++ Source/SteamNotifierHelper/App.config
