+++ Source/SteamNotifierHelper/Helper.resx
