+++ Source/SteamNotifierHelper/Program.cs
@@ -4,7 +4,7 @@ using System.Linq;
 using System.Threading.Tasks;
 using System.Windows.Forms;
 
-namespace SN_Utility {
+namespace SteamNotifierHelper {
     static class Program {
         /// <summary>
         /// The main entry point for the application.
@@ -14,7 +14,7 @@ namespace SN_Utility {
         {
             Application.EnableVisualStyles();
             Application.SetCompatibleTextRenderingDefault(false);
-            Application.Run(new SNUtility());
+            Application.Run(new Helper());
         }
     }
 }
