+++ Source/SteamNotifierHelper/Properties/AssemblyInfo.cs
