+++ Source/SteamNotifierHelper/Properties/Resources.resx
