+++ Source/SteamNotifierHelper/Properties/Settings.settings
