+++ Source/SteamNotifierHelper/Resources/icon-bg.ico
