+++ Source/SteamNotifierHelper/Resources/icon-bg.png
