+++ Source/SteamNotifierHelper/Resources/icon-bg2.png
