+++ Source/SteamNotifierHelper/Resources/icon.ico
