+++ Source/SteamNotifierHelper/Resources/icon.png
